Fixed a warning with the last commit and added 64 bit support to the transfer function to support 64 bit discord ids. Also fixed a bug with the transfer function where we never actually opened the previous accounts wallet.

+++ XMR Discord TipBot/Account.h
@@ -16,7 +16,7 @@ public:
 	unsigned long long		getUnlockedBalance() const;
 	const std::string &		getMyAddress() const;
 
-	TransferRet				transferMoneytoAnotherDiscordUser(unsigned long long amount, unsigned int Discord_ID) const;
+	TransferRet				transferMoneytoAnotherDiscordUser(unsigned long long amount, unsigned long long Discord_ID) const;
 	TransferRet				transferMoneyToAddress(unsigned long long amount, const std::string & address) const;
 private:
 	RPC						RPCServ;
